--- src/fs-operations/index.js
@@ -1,8 +1,0 @@
-import { readFileToConsole } from './readFile.js';
-import { createFile } from './createFile.js';
-import { renameFile } from './renameFile.js';
-import { copyFile } from './copyFile.js';
-import { deleteFile } from './deleteFile.js';
-import { moveFile } from './moveFile.js';
-
-export { readFileToConsole, createFile, renameFile, copyFile, deleteFile, moveFile };
