+++ src/index.js
@@ -1,8 +1,8 @@
 import { pipeline } from 'node:stream/promises';
 
-import { printUserGreetings, printUserGoodbye } from './user/user.js';
-import { directoryController } from './directory/directory.controller.js';
-import { controller } from './controller/controller.js';
+import { printUserGreetings, printUserGoodbye } from './modules/user/user.js';
+import { directoryController } from './modules/directory/directory.controller.js';
+import { controller } from './controller/main.controller.js';
 
 printUserGreetings();
 directoryController.printCurrentDirectory();
