+++ src/modules/archive/decompress.js
@@ -1,26 +1,28 @@
 import { open as fsPromisesOpen } from 'node:fs/promises';
 import { pipeline } from 'node:stream/promises';
-import path from 'node:path';
+import { createBrotliDecompress } from 'node:zlib';
 
-import { getAbsolutePath } from '../helpers/helper.js';
+import { getAbsolutePath } from '../../common/helpers/helper.js';
 
-export const copyFile = async (sourceFilePath, destinationDirPath, currentDirectory) => {
+export const decompress = async (sourceFilePath, destinationFilePath, currentDirectory) => {
   let sourceFh, destinationFh;
-  
+
   try {
     const absoluteSourceFilePath = getAbsolutePath(sourceFilePath, currentDirectory);
     sourceFh = await fsPromisesOpen(absoluteSourceFilePath);
     const readStream = await sourceFh.createReadStream();
 
-    const absoluteDestinationFilePath = getAbsolutePath(path.join(destinationDirPath, path.basename(sourceFilePath)), currentDirectory);
+    const absoluteDestinationFilePath = getAbsolutePath(destinationFilePath, currentDirectory);
     destinationFh = await fsPromisesOpen(absoluteDestinationFilePath, 'wx');
     const writeStream = await destinationFh.createWriteStream();
 
-    await pipeline(readStream, writeStream);
+    const brotliDecompressStream = createBrotliDecompress();
+
+    await pipeline(readStream, brotliDecompressStream, writeStream);
   } catch(error) {
     throw new Error(error);
   } finally {
     sourceFh && sourceFh.close();
     destinationFh && destinationFh.close();
   }
-}
+};
