+++ src/modules/fs-operations/createFile/createFile.js
@@ -1,7 +1,7 @@
 import path from 'node:path';
 import { open as fsPromisesOpen } from 'node:fs/promises';
 
-import { checkFileName } from '../helpers/helper.js';
+import { checkFileName } from '../../../common/helpers/helper.js';
 
 export const createFile = async (fileName, currentDirectory) => {
   checkFileName(fileName);
