+++ src/modules/fs-operations/deleteFile/deleteFile.js
@@ -1,6 +1,6 @@
 import { rm as fsPromisesRemove } from 'node:fs/promises';
 
-import { getAbsolutePath, checkIsFile} from '../helpers/helper.js';
+import { getAbsolutePath, checkIsFile} from '../../../common/helpers/helper.js';
 
 export const deleteFile = async (filePath, currentDirectory) => {
   const absoluteFilePath = getAbsolutePath(filePath, currentDirectory);
