+++ src/modules/fs-operations/index.js
@@ -1,0 +1,8 @@
+import { readFileToConsole } from './readFile/readFile.js';
+import { createFile } from './createFile/createFile.js';
+import { renameFile } from './renameFile/renameFile.js';
+import { copyFile } from './copyFile/copyFile.js';
+import { deleteFile } from './deleteFile/deleteFile.js';
+import { moveFile } from './moveFile/moveFile.js';
+
+export { readFileToConsole, createFile, renameFile, copyFile, deleteFile, moveFile };
