+++ src/modules/fs-operations/moveFile/moveFile.js
@@ -1,4 +1,4 @@
-import { copyFile, deleteFile } from './index.js';
+import { copyFile, deleteFile } from '../index.js';
 
 export const moveFile = async (sourceFilePath, destinationDirPath, currentDirectory) => {
   await copyFile(sourceFilePath, destinationDirPath, currentDirectory);
